Removed LDAP
I'm not sure if we're removing LDAP or not but a lot of people are
having major issues with authentication because they can't sign in/
sign up on the LDAP one. Seems to be not working. An idea was being
floated around that we're switching to a regular authentication flow.

If we are, this commit reinstalls Devise (without all the LDAP stuff).
So, now you have a sign up page....also removed the session store because
Devise's <code>Trackable</code> module does the same thing.