CSS Styling Caused Form Error
There was an extra form tag that was causing the wrong form to be
submitted. For some reason, two forms were nested, one within
another which caused only the parent to submit while the child never
did. The parent was an empty <form> tag with a class while the child
was the Devise form.