Refactored Sir Trevor Backend Code To Handle Updating Articles
Previously, we couldn't update articles. Now we can. This basically
deletes the previous article and updates it with the new block. Still,
have to handle images slightly more effectively but it will allow a user
to overwrite their previous content with new content.

Also fixed someone's javascript that didn't use the proper on ready handler.