Fix Article Creation and Image Upload Bugs, Implement Eagerloading, Remove Turbolinks
The refactoring done yesterday to allow a user to edit a block broke
the creation of articles. Additionally, the image upload was changed
to prevent reuploading images. Both had some small bugs that should
be fixed.

Eagerloading was a major problem with this app which is
understandable as it's more of an advanced Rails topic but basically
when you call Category.all then call subcategories on each category,
that creates a whole bunch of queries (one for each subcategory).
Now that eagerloading is implemented, it should do the minimum necessary
(usually 1-2). This was done with the gem 'goldilocks'. I would preferably
like to do this by hand but the gem does it automatically and will cover
any additional cases. This way we don't have to worry about it.

Turbolinks was causing issues with SirTrevor so it got the boot. However,
the speed improvements from eagerloading make it unnoticiable.

I also commented out the equation text block.