Fix error with admin panel equation text dashboard, allow users to be created in dashboard, add sign out to admin panel, remove sign up ability
Fixed an error with the admin panel but creating the missing controller and dashboard. Wasn't auto generated.
Created a virtual password and password confirmation field to the user dashboard. Creation and login with that user works.
Added a sign out URL to the admin panel, this needs to be styled? Trying to figure that out.
Removed the abilty to sign up and the sign up URL, not needed. Only admin accounts should be created.