Revert "minify the css sheet"
This commit broke the website but that is not the reason for the revert.   Minify CSS removes all comments and long variable names which is awesome for a large project where we are concerned with every ms we can cut from load time.  This is not the case for us.  We should be added more comments and documenting our code MORE not stripping away documentation to save a few BYTES of data.

This reverts commit c4ce1bdb62608906e8c3ff1825cfdbdeebb0d4a8.