Fixed Image Upload Error on Edit
There was an issue with Sir Trevor where it wasn't submitting the id
of the image on re-upload. This is because on the inital upload, it takes
the response but on re-uploads, it kinda submits a mix of my response
and its default response.

In order to get around this, there's now a check if the id was provided -
that is, the intial article creation. If not, then we use the file path
and find the image. We could probably just use the filepath but the id
is much faster.